完成商家添加商品模块，搭建了阿里服务器的redis和ftp，将上传的图片储存进ftp。

+++ src/main/java/com/zjh/e/service/CommodityTypeService.java
@@ -9,6 +9,6 @@ import java.util.List;
  */
 public interface CommodityTypeService extends BaseService<CommodityType> {
 
-     List<CommodityType> selectByExample(long parentId);
+     List<CommodityType> selectByExample(Integer parentId);
 
 }
